Extract functionality into a library
As is traditional, it's time for integration tests
