+++ src/katex.rs
@@ -1,0 +1,61 @@
+use crate::{write, EXPORT_DIR};
+use anyhow::{bail, Result};
+use futures_util::stream::{FuturesUnordered, TryStreamExt};
+use reqwest::Client;
+use std::path::Path;
+use tokio::task::JoinHandle;
+
+pub fn download(client: Client) -> JoinHandle<Result<()>> {
+    const CDN_URL: &str = "https://cdn.jsdelivr.net/npm/katex@0.15.1/dist/";
+    const KATEX_DIR: &str = "katex";
+
+    async fn download_file(client: &Client, file: &str) -> Result<()> {
+        let response = client.get(format!("{}{}", CDN_URL, file)).send().await?;
+
+        let status = response.status();
+        if status.is_client_error() || status.is_server_error() {
+            bail!(
+                "Download request for file {} failed with status code {}",
+                file,
+                status
+            )
+        }
+
+        let bytes = response.bytes().await?;
+
+        write(Path::new(EXPORT_DIR).join(KATEX_DIR).join(file), bytes).await?;
+
+        Ok(())
+    }
+
+    tokio::spawn(async move {
+        let response = client
+            .get(format!("{}{}", CDN_URL, "katex.min.css"))
+            .send()
+            .await?;
+
+        let katex_styles = response.text().await?;
+
+        let assets_downloads = katex_styles
+            .split("url(")
+            .skip(1)
+            .map(|part| part.split(')').next())
+            .map(|file| {
+                file.ok_or_else(|| {
+                    anyhow::format_err!("Failed to parse asset URL from Katex stylesheet")
+                })
+            })
+            .map(|result| result.map(|file| download_file(&client, file)))
+            .collect::<Result<FuturesUnordered<_>>>()?;
+
+        tokio::try_join!(
+            write(
+                Path::new(EXPORT_DIR).join(KATEX_DIR).join("katex.min.css"),
+                &katex_styles
+            ),
+            assets_downloads.try_collect::<()>(),
+        )?;
+
+        Ok(())
+    })
+}
